Parser fix, Nav bar points to correct pages, fix uploading to database

+++ ArkansasAssetBuilders/Pages/Clients/Edit.cshtml.cs
@@ -56,7 +56,7 @@ namespace ArkansasAssetBuilders.Pages.Clients
             }
             catch (DbUpdateConcurrencyException)
             {
-                if (!ClientExists(Client.ID))
+                if (!ClientExists((int)Client.ID))
                 {
                     return NotFound();
                 }
